[client] | refactor(svgs): 개별 svgs 임포트 방식 수정
svgs 에 index.tsx를 추가하였으며 개별 SVG 임포트 방식을 변경하였습니다.
