+++ packages/client/svgs/index.tsx
@@ -1,0 +1,23 @@
+import ArrowDropDown from './ArrowDropDown';
+import CloudSVG from './CloudSVG';
+import FavoriteSVG from './FavoriteSVG';
+import GitHubSVG from './GitHubSVG';
+import HotlistSVG from './HotlistSVG';
+import LatestSVG from './LatestSVG';
+import LogoSVG from './LogoSVG';
+import ProfileSVG from './ProfileSVG';
+import SearchSVG from './SearchSVG';
+import TopicsSVG from './TopicsSVG';
+
+export {
+  ArrowDropDown,
+  CloudSVG,
+  FavoriteSVG,
+  GitHubSVG,
+  HotlistSVG,
+  LatestSVG,
+  LogoSVG,
+  ProfileSVG,
+  SearchSVG,
+  TopicsSVG,
+};
